Refactor : 핀 정리하기 - pin-organize.js 로 파일 분리해놨던것 pin.js로 합침
