+++ controllers/pins.js
@@ -40,3 +40,16 @@ export const savePin = async (req, res) => {
     res.status(error.statusCode || 500).json({ message: error.message });
   }
 };
+
+export const organizePins = async (req, res) => {
+  try {
+    const userId = req.userId;
+    const pinId = req.body.pin_id;
+    const boardId = req.body.board_id;
+
+    await pinsService.organizePins(userId, pinId, boardId);
+    res.status(201).json({ message: 'SUCCESS' });
+  } catch (error) {
+    res.status(error.statusCode || 500).json({ message: error.message });
+  }
+};
