--- services/pin-organize.js
@@ -1,5 +1,0 @@
-import * as pinsOrganizeModels from '../models/pin-organize.js';
-
-export async function organizePins(userId, pinId, boardId) {
-  await pinsOrganizeModels.organizePins(userId, pinId, boardId);
-}
